tests: split official ONNX doc generation
